Fix input for multi-valued attributes
If the user dynamically adds a multi-valued attribute using Add Attribute, it is not recognized as a multi-valued input field back on the Attribute Statements form. This fix adds the bootstrap tags input markup needed so that the user can correctly edit all multi-valued attributes, not just the initial Groups attribute.